Refactor: ヘッダーコンポーネントをfeatures/landing-page配下に移動
- src/components/header.tsx → src/features/landing-page/components/header.tsx
- src/components/language-switcher.tsx → src/features/landing-page/components/language-switcher.tsx
- page.tsxのインポートパスを更新
- 機能ベースのディレクトリ構造に最適化
